requirements.txt
numpy==1.14.2
torch==0.4.1
tqdm
python-Levenshtein